clean and restructure DI
