+++ src/backend/ReUse.API/DependencyInjection.cs
@@ -1,14 +1,33 @@
-﻿using System.Text.Json.Serialization;
+﻿using System.Text;
+using System.Text.Json.Serialization;
 
 using FluentValidation;
 using FluentValidation.AspNetCore;
 
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 
+using Reuse.Infrastructure.Identity.Models;
+
+using ReUse.API.Responses;
 using ReUse.Application.DTOs.Users.UserProfile;
+using ReUse.Application.Errors;
+using ReUse.Application.Interfaces.Services.External;
+using ReUse.Application.Options;
 using ReUse.Infrastructure.Identity;
+using ReUse.Infrastructure.Interfaces.Repositories;
+using ReUse.Infrastructure.Interfaces.Services;
 using ReUse.Infrastructure.Persistence;
+using ReUse.Infrastructure.Repositories;
+using ReUse.Infrastructure.Security.Authorization;
+using ReUse.Infrastructure.Services.Auth;
+using ReUse.Infrastructure.Services.Caching;
+using ReUse.Infrastructure.Services.Communication;
+using ReUse.Infrastructure.Services.Identity;
 
 namespace ReUse.API;
 
@@ -80,21 +99,145 @@ public static class DependencyInjection
         return services;
     }
 
-    public static IServiceCollection AddDatabase(this IServiceCollection services,
-        IConfiguration configuration)
+    #region Auth
+    public static IServiceCollection AddAuth(this IServiceCollection services, IConfiguration configuration)
     {
-        var connectionString = configuration.GetConnectionString("pgsql");
+        var jwtOptions = configuration.GetSection("Jwt").Get<JwtOptions>()
+                         ?? throw new InvalidOperationException("JWT configuration is missing");
 
-        services.AddDbContext<ApplicationDbContext>(options =>
-            options.UseNpgsql(connectionString,
-                b => b.MigrationsAssembly(typeof(ApplicationDbContext).Assembly.FullName)));
+        services.Configure<JwtOptions>(configuration.GetSection("Jwt"));
+        services.Configure<RefreshTokenOptions>(configuration.GetSection("RefreshToken"));
+        services.Configure<EmailOptions>(configuration.GetSection("EmailSettings"));
 
-        services.AddDbContext<AppIdentityDbContext>(options =>
-            options.UseNpgsql(connectionString,
-                b => b.MigrationsAssembly(typeof(AppIdentityDbContext).Assembly.FullName)));
+        services.AddAuthentication(options =>
+        {
+            options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+            options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+        })
+        .AddJwtBearer(options =>
+        {
+            options.RequireHttpsMetadata = false;
+            options.SaveToken = true;
+
+            options.TokenValidationParameters = new TokenValidationParameters
+            {
+                ValidateIssuer = true,
+                ValidateAudience = true,
+                ValidateLifetime = true,
+                ValidateIssuerSigningKey = true,
+                ValidIssuer = jwtOptions.Issuer,
+                ValidAudience = jwtOptions.Audience,
+                IssuerSigningKey = new SymmetricSecurityKey(
+                    Encoding.UTF8.GetBytes(jwtOptions.SigningKey))
+            };
+
+            options.Events = new JwtBearerEvents
+            {
+                OnMessageReceived = context =>
+                {
+                    var token = context.Request.Cookies["access_token"];
+                    if (!string.IsNullOrEmpty(token))
+                        context.Token = token;
+
+                    return Task.CompletedTask;
+                },
+                OnChallenge = context =>
+                {
+                    context.HandleResponse();
+
+                    context.Response.StatusCode = StatusCodes.Status401Unauthorized;
+                    context.Response.ContentType = "application/json";
+
+                    return context.Response.WriteAsJsonAsync(new ErrorResponse
+                    {
+                        Message = "User is not authenticated.",
+                        Code = ErrorsCode.Unauthorized
+                    });
+                },
+                OnForbidden = context =>
+                {
+                    context.Response.StatusCode = StatusCodes.Status403Forbidden;
+                    context.Response.ContentType = "application/json";
+
+                    return context.Response.WriteAsJsonAsync(new ErrorResponse
+                    {
+                        Message = "User is not authorized to access this resource.",
+                        Code = ErrorsCode.Forbidden
+                    });
+                }
+            };
+        });
+
+        services.AddAuthorization(options =>
+        {
+            options.FallbackPolicy = new AuthorizationPolicyBuilder()
+                .RequireAuthenticatedUser()
+                .AddRequirements(new ActiveUserRequirement())
+                .Build();
+        });
 
         return services;
     }
+    #endregion
+
+    #region Identity
+    public static IServiceCollection AddIdentityServices(this IServiceCollection services)
+    {
+        services.AddIdentityCore<ApplicationUser>(options =>
+        {
+            options.Password.RequiredLength = 8;
+            options.Password.RequireDigit = true;
+            options.Password.RequireLowercase = true;
+            options.Password.RequireUppercase = true;
+            options.Password.RequireNonAlphanumeric = true;
+
+            options.User.RequireUniqueEmail = true;
+
+            options.User.AllowedUserNameCharacters =
+                "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_";
+
+            options.SignIn.RequireConfirmedEmail = true;
+            options.SignIn.RequireConfirmedPhoneNumber = true;
+        })
+        .AddRoles<IdentityRole>()
+        .AddEntityFrameworkStores<AppIdentityDbContext>()
+        .AddDefaultTokenProviders();
+
+        return services;
+    }
+
+    #endregion
+
+    #region Http
+    public static IServiceCollection AddHttpServices(this IServiceCollection services)
+    {
+        services.AddSingleton<IAuthorizationMiddlewareResultHandler, AuthorizationResultHandler>();
+
+        return services;
+    }
+
+    #endregion
+
+    #region CorsPolicy
+    public static IServiceCollection AddCorsPolicy(this IServiceCollection services)
+    {
+        services.AddCors(options =>
+        {
+            options.AddPolicy("AllowReactApp",
+                policy =>
+                {
+                    policy.WithOrigins("http://localhost:5173")
+                        .AllowAnyHeader()
+                        .AllowAnyMethod()
+                        .AllowCredentials();
+                });
+        });
+
+        return services;
+    }
+
+    #endregion
+
     #region Register FluentValidation validators
     public static IServiceCollection AddValidation(this IServiceCollection services)
     {
@@ -106,7 +249,7 @@ public static class DependencyInjection
     }
     #endregion
 
-    #region Register AutoMapper profiles
+    #region AutoMapper
     public static IServiceCollection AddAutoMapperProfiles(this IServiceCollection services)
     {
         services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
@@ -114,4 +257,5 @@ public static class DependencyInjection
         return services;
     }
     #endregion
+
 }
