+++ src/backend/ReUse.API/Program.cs
@@ -1,41 +1,6 @@
-using System;
-using System.Text;
-using System.Text.Json.Serialization;
-
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.OpenApi.Models;
-
-using Reuse.Infrastructure.Identity.Models;
-
-using ReUse.API.Middlewares;
-using ReUse.API.Responses;
+using ReUse.API.Extensions;
 using ReUse.Application;
-using ReUse.Application.Errors;
-using ReUse.Application.Interfaces;
-using ReUse.Application.Interfaces.Services.External;
-using ReUse.Application.Mappers;
-using ReUse.Application.Options;
 using ReUse.Infrastructure;
-using ReUse.Infrastructure.Identity;
-using ReUse.Infrastructure.Interfaces.Repositories;
-using ReUse.Infrastructure.Interfaces.Services;
-using ReUse.Infrastructure.Persistence;
-using ReUse.Infrastructure.Repositories;
-using ReUse.Infrastructure.Security.Authorization;
-using ReUse.Infrastructure.Seeders;
-using ReUse.Infrastructure.Services.Auth;
-using ReUse.Infrastructure.Services.Caching;
-using ReUse.Infrastructure.Services.Communication;
-using ReUse.Infrastructure.Services.Identity;
-using ReUse.Infrastructure.UnitOfWork;
 
 using Serilog;
 
@@ -47,186 +12,42 @@ public class Program
     {
         var builder = WebApplication.CreateBuilder(args);
 
+        #region Layers
         builder.Services.AddPresentation();
-        builder.Services.AddSwagger();
-        builder.Services.AddAutoMapperProfiles();
-        builder.Services.AddDatabase(builder.Configuration);
         builder.Services.AddApplication(builder.Configuration);
         builder.Services.AddInfrastructure(builder.Configuration);
+        #endregion
+
+        #region Cross-cutting
+        builder.Services.AddHttpServices();
+        builder.Services.AddIdentityServices();
+        builder.Services.AddAuth(builder.Configuration);
         builder.Services.AddValidation();
+        builder.Services.AddAutoMapperProfiles();
+        builder.Services.AddCorsPolicy();
+        #endregion
 
+        // Add Swagger
 
-
+        builder.Services.AddSwagger();
 
         builder.Host.UseSerilog((context, configuration) =>
-            configuration.ReadFrom.Configuration(context.Configuration));
-
-        var jwtOptions = builder.Configuration.GetSection("Jwt").Get<JwtOptions>()
-                         ?? throw new InvalidOperationException("JWT configuration is missing");
-        builder.Services.Configure<JwtOptions>(
-            builder.Configuration.GetSection("Jwt"));
-
-        builder.Services.Configure<RefreshTokenOptions>(
-            builder.Configuration.GetSection("RefreshToken"));
-
-        builder.Services.Configure<EmailOptions>(
-            builder.Configuration.GetSection("EmailSettings"));
-
-        // Identity
-        builder.Services.AddIdentityCore<ApplicationUser>(options =>
-                {
-                    options.Password.RequiredLength = 8;
-                    options.Password.RequireDigit = true;
-                    options.Password.RequireLowercase = true;
-                    options.Password.RequireUppercase = true;
-                    options.Password.RequireNonAlphanumeric = true;
-
-                    options.User.RequireUniqueEmail = true;
-
-                    options.User.AllowedUserNameCharacters =
-                        "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_";
-
-                    options.SignIn.RequireConfirmedEmail = true;
-                    options.SignIn.RequireConfirmedPhoneNumber = true;
-                }
-            )
-            .AddRoles<IdentityRole>()
-            .AddEntityFrameworkStores<AppIdentityDbContext>()
-            .AddDefaultTokenProviders();
-
-        // Auth
-        builder.Services.AddAuthentication(options =>
-            {
-                options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-            .AddJwtBearer(JwtBearerDefaults.AuthenticationScheme,
-                options =>
-                {
-                    options.RequireHttpsMetadata = false; // to dev env
-                    options.SaveToken = true;
-                    options.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidateIssuer = true,
-                        ValidateAudience = true,
-                        ValidateLifetime = true,
-                        ValidateIssuerSigningKey = true,
-                        ValidIssuer = jwtOptions.Issuer,
-                        ValidAudience = jwtOptions.Audience,
-                        IssuerSigningKey = new SymmetricSecurityKey(
-                            Encoding.UTF8.GetBytes(jwtOptions.SigningKey)
-                        ),
-                    };
-                    options.Events = new JwtBearerEvents
-                    {
-                        OnMessageReceived = context =>
-                        {
-                            // read token from cookie instead of header
-                            var token = context.Request.Cookies["access_token"];
-
-                            if (!string.IsNullOrEmpty(token))
-                            {
-                                context.Token = token;
-                            }
-
-                            return Task.CompletedTask;
-                        },
-                        OnChallenge = context =>
-                        {
-                            context.HandleResponse();
-
-                            context.Response.StatusCode = StatusCodes.Status401Unauthorized;
-                            context.Response.ContentType = "application/json";
-
-                            var response = new ErrorResponse
-                            {
-                                Message = "User is not authenticated.",
-                                Code = ErrorsCode.Unauthorized
-                            };
-
-                            return context.Response.WriteAsJsonAsync(response);
-                        },
-
-                        OnForbidden = context =>
-                        {
-                            context.Response.StatusCode = StatusCodes.Status403Forbidden;
-                            context.Response.ContentType = "application/json";
-
-                            var response = new ErrorResponse
-                            {
-                                Message = "User is not authorized to access this resource.",
-                                Code = ErrorsCode.Forbidden
-                            };
-
-                            return context.Response.WriteAsJsonAsync(response);
-                        }
-                    };
-                }
-            );
-
-        builder.Services.AddAuthorization(options =>
-        {
-            options.FallbackPolicy = new AuthorizationPolicyBuilder()
-                .RequireAuthenticatedUser()
-                .AddRequirements(new ActiveUserRequirement())
-                .Build();
-        });
-        builder.Services.AddSingleton<IAuthorizationMiddlewareResultHandler, AuthorizationResultHandler>();
-        builder.Services.AddScoped<ITokenService, TokenService>();
-        builder.Services.AddScoped<IEmailConfirmationService, EmailConfirmationService>();
-        builder.Services.AddScoped<IPasswordResetService, PasswordResetService>();
-        builder.Services.AddScoped<IEmailService, EmailService>();
-        builder.Services.AddScoped<IOtpService, OtpService>();
-        builder.Services.AddScoped<IIdentityUserRepository, IdentityUserRepository>();
-
-        builder.Services.AddMemoryCache();
-        builder.Services.AddSingleton<IAppCache, MemoryCacheService>();
-
-        builder.Services.AddCors(options =>
-        {
-            options.AddPolicy("AllowReactApp",
-                policy =>
-                {
-                    policy.WithOrigins("http://localhost:5173")
-                        .AllowAnyHeader()
-                        .AllowAnyMethod()
-                        .AllowCredentials(); // important if using cookies/auth
-                });
-        });
+        configuration.ReadFrom.Configuration(context.Configuration));
 
         var app = builder.Build();
 
-        // Apply pending migrations automatically
-        using (var scope = app.Services.CreateScope())
-        {
-            var identity = scope.ServiceProvider.GetRequiredService<AppIdentityDbContext>();
-            var db = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-            identity.Database.Migrate();
-            db.Database.Migrate();
-        }
+        app.ApplyMigrations();
 
-        // Seed Data
-        using (var scope = app.Services.CreateScope())
-        {
-            var services = scope.ServiceProvider;
-            RoleSeeder.SeedRolesAsync(services).GetAwaiter().GetResult();
-            IdentitySeeder.SeedAdminAsync(services).GetAwaiter().GetResult();
-        }
+
+        app.SeedData();
 
         if (app.Environment.IsDevelopment())
         {
-            app.MapOpenApi();
-            app.UseSwagger();
-            app.UseSwaggerUI();
+            app.UseSwaggerServices();
         }
 
-        app.UseSerilogRequestLogging();
-        app.UseHttpsRedirection();
-        app.UseCors("AllowReactApp");
-        app.UseMiddleware<ExceptionMiddleware>();
-        app.UseAuthentication();
-        app.UseAuthorization();
-        app.MapControllers();
+        app.UsePipeline();
+
         app.Run();
     }
 }
